SSG-3695: WS-Policy and metrics are now deleted from UDDI when a gateway service is unpublished.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@25251 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/gateway/server/src/main/java/com/l7tech/server/uddi/UDDICoordinator.java
@@ -503,20 +503,36 @@ public class UDDICoordinator implements ApplicationContextAware, ApplicationList
                     origService.getUddiServiceKey() ), null );
         }
 
+        final Map<Long,WsPolicyUDDIEvent> wsPolicyEventMap = new HashMap<Long,WsPolicyUDDIEvent>(); // one event per registry
+
         // Delete stale entries
         final Set<Triple<Long,Long,String>> registryServiceKeys = registryServiceMap.keySet();
         for ( UDDIBusinessServiceStatus serviceStatus : status ) {
             Triple<Long,Long,String> key = new Triple<Long,Long,String>( serviceStatus.getUddiRegistryOid(), serviceStatus.getPublishedServiceOid(), serviceStatus.getUddiServiceKey() );
             if ( !registryServiceKeys.contains(key) ) {
-                logger.info( "Deleting business service status for " + key + "." );
-                uddiBusinessServiceStatusManager.delete( serviceStatus );
+                if ( serviceStatus.getUddiMetricsReferenceStatus() == UDDIBusinessServiceStatus.Status.NONE &&
+                     serviceStatus.getUddiPolicyStatus() == UDDIBusinessServiceStatus.Status.NONE ) {
+                    logger.info( "Deleting business service status for " + key + "." );
+                    uddiBusinessServiceStatusManager.delete( serviceStatus );
+                } else {
+                    registryServiceMap.put( key, serviceStatus );
+                    updateUDDIBusinessServiceStatus(
+                            registryServiceMap,
+                            wsPolicyEventMap,
+                            serviceStatus.getPublishedServiceOid(),
+                            serviceStatus.getUddiRegistryOid(),
+                            serviceStatus.getUddiServiceKey(),
+                            serviceStatus.getUddiServiceName(),
+                            false,
+                            false,
+                            null );
+                }
             } else {
                 registryServiceMap.put( key, serviceStatus );
             }
         }
 
         // Update / create service status
-        Map<Long,WsPolicyUDDIEvent> wsPolicyEventMap = new HashMap<Long,WsPolicyUDDIEvent>(); // one event per registry
         for ( UDDIProxiedServiceInfo proxiedServiceInfo : proxiedServices ) {
             if ( proxiedServiceInfo.getProxiedServices() != null ) {
                 final long publishedServiceOid = proxiedServiceInfo.getPublishedServiceOid();
